Restructure MVVM modules and add feed data manager
