+++ ViewModels/VideoTabViewModel.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+final class VideoTabViewModel: ObservableObject {
+    @Published var videos: [VideoItem]
+
+    private let dataManager: FeedDataManager
+
+    init(dataManager: FeedDataManager = .shared) {
+        self.dataManager = dataManager
+        loadVideos()
+    }
+
+    private func loadVideos() {
+        videos = dataManager.videoItems
+    }
+}
