+++ Views/Components/Ads/AdsPostView.swift
